refactor: moved pentagon domains into separate folder
