+++ src/abstract-interpretation/pentagon/closed-pentagon-domain.ts
