+++ src/abstract-interpretation/pentagon/upper-bounds-domain.ts
